// apps/vendor/src/services/models/StudioViewPagedCollectionStandardResponse.ts
/* istanbul ignore file */
/* tslint:disable */
/* eslint-disable */

import type { HttpStatusCode } from './HttpStatusCode';
import type { StudioViewPagedCollection } from './StudioViewPagedCollection';

export type StudioViewPagedCollectionStandardResponse = {
    href?: string | null;
    relations?: Array<string> | null;
    method?: string | null;
    routeName?: string | null;
    routeValues?: any;
    status?: boolean;
    message?: string | null;
    data?: StudioViewPagedCollection;
    statusCode?: HttpStatusCode;
    errors?: any;
};