// apps/vendor/src/services/models/StudioViewStandardResponse.ts
/* istanbul ignore file */
/* tslint:disable */
/* eslint-disable */

import type { HttpStatusCode } from './HttpStatusCode';
import type { StudioView } from './StudioView';

export type StudioViewStandardResponse = {
    href?: string | null;
    relations?: Array<string> | null;
    method?: string | null;
    routeName?: string | null;
    routeValues?: any;
    status?: boolean;
    message?: string | null;
    data?: StudioView;
    statusCode?: HttpStatusCode;
    errors?: any;
};